feat:  crud event

+++ lib/resources/firestore/event_collection.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 
+import 'package:chanthaburi_app/models/events/event_model.dart';
 import 'package:chanthaburi_app/resources/firebase_storage.dart';
 import 'package:chanthaburi_app/utils/my_constant.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
@@ -11,28 +12,49 @@ final CollectionReference eventCollection =
     _firestore.collection(MyConstant.eventCollection);
 
 class EventCollection {
-  static Stream<QuerySnapshot> events() {
-    Stream<QuerySnapshot> _events = eventCollection.snapshots();
+  static Stream<QuerySnapshot<EventModel>> events() {
+    Stream<QuerySnapshot<EventModel>> _events = eventCollection
+        .withConverter<EventModel>(
+            fromFirestore: (snapshot, _) =>
+                EventModel.fromMap(snapshot.data()!),
+            toFirestore: (map, _) => map.toMap())
+        .snapshots();
     return _events;
   }
 
-  static Future<Map<String, dynamic>> createEvent(File eventRef) async {
+  static Future<Map<String, dynamic>> upsertEvent(
+      EventModel eventModel, File? eventRef, String? docId) async {
     try {
-      String fileName = basename(eventRef.path);
-      String imageURL =
-          await StorageFirebase.uploadImage("images/event/$fileName", eventRef);
-      await eventCollection.add({"eventURL": imageURL});
-      return {"status": "200", "message": "สร้างกิจกรรมเรียบร้อย"};
+      if (eventRef != null) {
+        if (eventModel.url.isNotEmpty) {
+          String referenceImage = StorageFirebase.getReference(eventModel.url);
+          StorageFirebase.deleteFile(referenceImage);
+        }
+        String fileName = basename(eventRef.path);
+        String imageURL = await StorageFirebase.uploadImage(
+            "images/event/$fileName", eventRef);
+        eventModel.url = imageURL;
+      }
+      if (docId == null) {
+        await eventCollection.add(eventModel.toMap());
+      } else {
+        await eventCollection.doc(docId).update(eventModel.toMap());
+      }
+
+      return {"status": "200", "message": "จัดการกิจกรรมเรียบร้อย"};
     } catch (e) {
-      return {"status": "400", "message": "สร้างกิจกรรมล้มเหลว"};
+      return {"status": "400", "message": "จัดการกิจกรรมล้มเหลว"};
     }
   }
 
   static Future<Map<String, dynamic>> deleteEvenet(
       String eventId, String imageUrl) async {
     try {
-      String referenceImage = StorageFirebase.getReference(imageUrl);
-      StorageFirebase.deleteFile(referenceImage);
+      if (imageUrl.isNotEmpty) {
+        String referenceImage = StorageFirebase.getReference(imageUrl);
+        StorageFirebase.deleteFile(referenceImage);
+      }
+
       await eventCollection.doc(eventId).delete();
       return {"status": "200", "message": "ลบกิจกรรมเรียบร้อย"};
     } catch (e) {
